refactor(stream): move streaming logic to dedicated module
* Move `StreamGenerateContent` and `RouteStream` types
* Move parsing logic for streamed responses
* Update imports in `client.rs` and `lib.rs`
* Adjust visibility of related fields/methods for inter-module access
