+++ src/lib.rs
@@ -48,14 +48,15 @@
 mod chat;
 mod client;
 mod error;
+mod stream;
 pub mod types;
 
 pub type Result<T> = std::result::Result<T, Error>;
 
 pub use chat::Chat;
 pub use client::Client;
-pub use client::{RouteStream, StreamGenerateContent};
 pub use error::Error;
+pub use stream::{RouteStream, StreamGenerateContent};
 
 /// Creates a new Gemini client instance using the default configuration.
 pub fn client() -> Client {
